Add Enum class to handle errors
I think it is a good idea to use an Enum class for errors, each plugin
will have to handle different kinds of errors and it is easyer if we all
use the same template.

I think it would be a good idea to implement a Response class that each plugin would have to return to handle errors.
Response:
 + error_code
 + error_message

and then define an error template that would display the error message.