Fix - Crash when no link was given
I handled the case when no link was given in the item of the xml file.

I also prevent that an publishing without title and description will be added to the rss_feed since PyRSS2Gen raise a TypeError if both title and description are None when creating a PyRSS2Gen.RSSItem. If this happens I reload the moderate_publishing page with an error message.