Add possibility to extend a template on a modular way
I added to moderate_post.html and new.html the possiblity for a plugin
to add easely some html to it.
To add specific information for your module in new or moderate_post you should create a folder in templates/plugins with the name of your plugin. Inside that folder you should create a file named moderate.html (or/and new.html if you want to add some stuff when creating a new post). Same if you want to add some personlized scripts. Make a file in the same directory (templates/plugins/plugin_name/) with the name moderate.js (or/and new.js).

If you want to add this feature to new templates you can juste add the
same modification that I added to new.html and moderate_post.html and
then in the plugin folder of your module juste add the file with the
corresponding name. Don't forget to make shure that you pass the channel
in argument when calling the template.

Co-authored-by: Maxmaw <max.mawait@gmail.com>
Co-authored-by: Sami <sami.basch@gmail.com>
Co-authored-by: bensim602 <benjamin.simon@student.uclouvain.be>
Co-authored-by: sluyters <arthur.sluyters@student.uclouvain.be>
Co-authored-by: Gui <guillaume.gheysen@stundent.uclouvain.be>
Co-authored-by: TGLuis <brutusltg@gmail.com>
Co-authored-by: tanguyl <tanguy.losseau@student.uclouvain.be>